requirements.txt
amqp==1.4.5
anyjson==0.3.3
apns==2.0.1
asn1crypto==0.24.0
billiard==3.3.0.16
bitarray==0.8.1
boto==2.39.0
celery==3.1.10
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
cssselect==0.9.1
Django==1.8.19  
django-allauth==0.29.0
django-bootstrap3==9.1.0
django-debug-toolbar==1.2.1
django-haystack==2.7.0
django-rest-swagger==0.1.14
django-ses==0.6.0
django-taggit===0.21.1
django-websocket-redis==0.4.1
djangorestframework==2.4.8
djangorestframework-jwt==0.1.4
docutils==0.11
drf-compound-fields==0.2.2
drip==1.0.8
elasticsearch==1.7.0
enum34==1.1.6
factory-boy==2.3.1
gevent==1.3.3
ghost==0.6.1
Ghost.py==0.2.3
greenlet==0.4.13
idna==2.7
ipaddress==1.0.22
kombu==3.0.30
lxml==3.3.6
MySQL-python==1.2.5
oauthlib==0.6.1
paho-mqtt==1.3.1
parse==1.8.4
Pillow==2.4.0
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
pyelasticsearch==1.4.1
PyJWT==0.2.1
pyOpenSSL==16.2.0
pysolr==3.2.0
python-dateutil==2.2
python-gcm==0.1.5
python-mpns==0.1.6
python-openid==2.2.5
pytz==2014.2
redis==2.9.1
requests==2.19.1
requests-oauthlib==0.4.0
rsa==3.1.4
shortuuid==0.4.2
simplejson==3.16.0
six==1.11.0
sqlparse==0.1.11
South==0.8.4
tinydb==3.9.0.post1
ua-parser==0.8.0
urllib3==1.23
user-agents==1.1.0
uWSGI==2.0.17
Whoosh==2.6.0
